requirements.txt
# Web Interface
gradio>=4.0.0

# Database
psycopg2-binary>=2.9.0
SQLAlchemy>=2.0.0
pgvector>=0.2.0

# Configuration
python-dotenv>=1.0.0

# LLM and Embeddings
google-generativeai>=0.3.0
